Remove comments about needing to do kref_get/kref_put in SG_IO code
The sg driver needs to do this to get a reference on the underlying
block device.  We do not need to do this because we implement SG_IO in
the block device ioctl handler itself, so we will already have taken
a reference via open() and will release it via close().